slink assures me that blobs can use size_t for ->len, and that this one <= 0 test is an oversight.

+++ include/vrt.h
@@ -178,7 +178,7 @@ struct strands {
 
 struct vrt_blob {
 	unsigned	type;
-	long		len;
+	size_t		len;
 	const void	*blob;
 };
